+++ Tests/KeyboardKitTests/Input/StandardKeyboardInputSetProviderTests.swift
@@ -15,7 +15,7 @@ class StandardKeyboardInputSetProviderTests: QuickSpec {
     
     override func spec() {
         
-        var provider: KeyboardInputSetProvider!
+        var provider: StandardKeyboardInputSetProvider!
         var context: MockKeyboardContext!
         
         var english: KeyboardInputSetProvider!
@@ -33,63 +33,60 @@ class StandardKeyboardInputSetProviderTests: QuickSpec {
             swedish = SwedishKeyboardInputSetProvider()
         }
         
-        describe("standard keyboard input set provider") {
+        describe("provider list") {
             
-            func alphabetic(for locale: String) -> KeyboardInputSet {
-                context.locale = Locale(identifier: locale)
-                return provider.alphabeticInputSet()
+            it("has locale-specific providers") {
+                let providers = provider.providerDictionary.dictionary
+                expect(providers.keys.count).to(equal(4))
+                expect(providers[LocaleKey.english.key] is EnglishKeyboardInputSetProvider).to(beTrue())
+                expect(providers[LocaleKey.german.key] is GermanKeyboardInputSetProvider).to(beTrue())
+                expect(providers[LocaleKey.italian.key] is ItalianKeyboardInputSetProvider).to(beTrue())
+                expect(providers[LocaleKey.swedish.key] is SwedishKeyboardInputSetProvider).to(beTrue())
             }
-            
-            func numeric(for locale: String) -> KeyboardInputSet {
-                context.locale = Locale(identifier: locale)
-                return provider.numericInputSet()
-            }
-            
-            func symbolic(for locale: String) -> KeyboardInputSet {
-                context.locale = Locale(identifier: locale)
-                return provider.symbolicInputSet()
-            }
+        }
+        
+        describe("input sets") {
             
             it("supports English") {
-                let locale = "en"
-                expect(alphabetic(for: locale)).to(equal(english.alphabeticInputSet()))
-                expect(numeric(for: locale)).to(equal(english.numericInputSet()))
-                expect(symbolic(for: locale)).to(equal(english.symbolicInputSet()))
+                context.locale = Locale(identifier: LocaleKey.english.key)
+                expect(provider.alphabeticInputSet()).to(equal(english.alphabeticInputSet()))
+                expect(provider.numericInputSet()).to(equal(english.numericInputSet()))
+                expect(provider.symbolicInputSet()).to(equal(english.symbolicInputSet()))
             }
             
             it("supports German") {
-                let locale = "de"
-                expect(alphabetic(for: locale)).to(equal(german.alphabeticInputSet()))
-                expect(numeric(for: locale)).to(equal(german.numericInputSet()))
-                expect(symbolic(for: locale)).to(equal(german.symbolicInputSet()))
+                context.locale = Locale(identifier: LocaleKey.german.key)
+                expect(provider.alphabeticInputSet()).to(equal(german.alphabeticInputSet()))
+                expect(provider.numericInputSet()).to(equal(german.numericInputSet()))
+                expect(provider.symbolicInputSet()).to(equal(german.symbolicInputSet()))
             }
             
             it("supports Italian") {
-                let locale = "it"
-                expect(alphabetic(for: locale)).to(equal(italian.alphabeticInputSet()))
-                expect(numeric(for: locale)).to(equal(italian.numericInputSet()))
-                expect(symbolic(for: locale)).to(equal(italian.symbolicInputSet()))
+                context.locale = Locale(identifier: LocaleKey.italian.key)
+                expect(provider.alphabeticInputSet()).to(equal(italian.alphabeticInputSet()))
+                expect(provider.numericInputSet()).to(equal(italian.numericInputSet()))
+                expect(provider.symbolicInputSet()).to(equal(italian.symbolicInputSet()))
             }
             
             it("supports Swedish") {
-                let locale = "sv"
-                expect(alphabetic(for: locale)).to(equal(swedish.alphabeticInputSet()))
-                expect(numeric(for: locale)).to(equal(swedish.numericInputSet()))
-                expect(symbolic(for: locale)).to(equal(swedish.symbolicInputSet()))
+                context.locale = Locale(identifier: LocaleKey.swedish.key)
+                expect(provider.alphabeticInputSet()).to(equal(swedish.alphabeticInputSet()))
+                expect(provider.numericInputSet()).to(equal(swedish.numericInputSet()))
+                expect(provider.symbolicInputSet()).to(equal(swedish.symbolicInputSet()))
             }
             
             it("has fallback support for specific locale") {
-                let locale = "en-US"
-                expect(alphabetic(for: locale)).to(equal(english.alphabeticInputSet()))
-                expect(numeric(for: locale)).to(equal(english.numericInputSet()))
-                expect(symbolic(for: locale)).to(equal(english.symbolicInputSet()))
+                context.locale = Locale(identifier: "en-US")
+                expect(provider.alphabeticInputSet()).to(equal(english.alphabeticInputSet()))
+                expect(provider.numericInputSet()).to(equal(english.numericInputSet()))
+                expect(provider.symbolicInputSet()).to(equal(english.symbolicInputSet()))
             }
             
-            it("has fallback support for non-existing locale") {
-                let locale = "abc"
-                expect(alphabetic(for: locale)).to(equal(english.alphabeticInputSet()))
-                expect(numeric(for: locale)).to(equal(english.numericInputSet()))
-                expect(symbolic(for: locale)).to(equal(english.symbolicInputSet()))
+            it("has fallback support for non-supported locale") {
+                context.locale = Locale(identifier: "es")
+                expect(provider.alphabeticInputSet()).to(equal(english.alphabeticInputSet()))
+                expect(provider.numericInputSet()).to(equal(english.numericInputSet()))
+                expect(provider.symbolicInputSet()).to(equal(english.symbolicInputSet()))
             }
         }
     }